@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <PropertyGroup>
-    <TargetFramework>net7.0-windows</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
     <Platforms>x64</Platforms>
     <RootNamespace>WhoSaidWhatNow</RootNamespace>
     <AssemblyName>WhoSaidWhatNow</AssemblyName>
@@ -34,8 +34,8 @@
 
   <ItemGroup>
     <PackageReference Include="Dalamud.ContextMenu" Version="1.3.1" />
-    <PackageReference Include="Dalamud.DrunkenToad" Version="1.8.4" />
-    <PackageReference Include="DalamudPackager" Version="2.1.12" />
+    <PackageReference Include="Dalamud.DrunkenToad" Version="1.9.1" />
+    <PackageReference Include="DalamudPackager" Version="2.1.13" />
     <PackageReference Include="System.Runtime" Version="4.3.1" />
     <Reference Include="FFXIVClientStructs">
       <HintPath>$(DalamudLibPath)FFXIVClientStructs.dll</HintPath>
