@@ -1,11 +1,9 @@
-using Dalamud.Game.Gui;
 using Dalamud.Game.Text;
 using Dalamud.Game.Text.SeStringHandling;
-using Dalamud.Game.Text.SeStringHandling.Payloads;
-using Dalamud.Logging;
 using Dalamud.Plugin.Services;
 using System;
 using WhoSaidWhatNow.Objects;
+using WhoSaidWhatNow.Utils;
 
 namespace WhoSaidWhatNow.Services
 {
@@ -32,14 +30,14 @@ namespace WhoSaidWhatNow.Services
             gui!.ChatMessage -= OnChatMessage;
         }
 
-        private void OnChatMessage(XivChatType type, uint senderId, ref SeString sender, ref SeString message, ref bool isHandled)
+        private void OnChatMessage(XivChatType type, int senderId, ref SeString sender, ref SeString message, ref bool isHandled)
         {
             if (Plugin.Config.Enabled == true)
             {
                 var senderName = sender.ToString();
                 //The basic ToString here includes any friends list icons and the server name, so we have to do Contains() for now
-                //this would be gross if it was messages but it should be okay given a person will probably only have 4-5 players tracked a time
-                var result = Plugin.Players.Find(x => senderName.Contains(x.Name));
+                //this would be gross if it was messages but it should be okay given a person will probably only have 4-5 players tracked a time <- written by a past idiot
+                var result = PlayerUtils.GetCurrentAndPlayers().Find(x => senderName.Contains(x.Name));
                 //PluginLog.Debug("onchat message triggered" + senderName);
 
                 if (result != null)
