@@ -132,11 +132,19 @@ public class TabIndividual
             }
             ImGui.EndChild();
 
-
+            ImGui.BeginChild(MainWindow.ID_PANEL_LEFT);
             //Reopen left window, populate selectable list
+            try
+            {
+                mainWindow.AddPlayerSelectable(Plugin.CurrentPlayer!);
+            }
+            catch
+            {
+
+            }
             foreach (var p in players)
             {
-                ImGui.BeginChild(MainWindow.ID_PANEL_LEFT);
+                
                 // catch NPEs silently
                 try
                 {
@@ -146,8 +154,9 @@ public class TabIndividual
                 {
 
                 }
-                ImGui.EndChild();
+                
             }
+            ImGui.EndChild();
 
             // Reopen right window, build the chat log
             // it's worth noting all of this stuff stays in memory and is only hidden when it's "closed"
@@ -157,7 +166,7 @@ public class TabIndividual
             if (Plugin.SelectedPlayer is not null)
             {
                 foreach (var c in from KeyValuePair<DateTime, ChatEntry> c in Plugin.ChatEntries
-                                  where Plugin.Config.ChannelToggles[c.Value.Type] == true && c.Value.Sender.Name.Contains(Plugin.SelectedPlayer.Name)
+                                  where Plugin.Config.ChannelToggles[c.Value.Type] && c.Value.Sender.Name.Contains(Plugin.SelectedPlayer.Name)
                                   select c)
                 {
                     ChatUtils.ShowMessage(c);
